Update all components in Calculator file
